Move license activation scripts to src/activacion
