+++ src/activacion/activacion.py
