+++ src/activacion/generador_licencia.py
